+++ scripts/redirects-generator.js
@@ -2,7 +2,8 @@
 const { default: axios } = require('axios');
 const fs = require('fs');
 const { isWhiteLabelAcademy, WHITE_LABEL_ACADEMY } = require('./_utils');
-const { getAsset } = require('./sitemap-generator/requests');
+// const { getAsset } = require('../src/utils/requests');
+const generatedRoutes = require('../src/utils/route-builder');
 require('dotenv').config({
   path: '.env.production',
 });
@@ -13,25 +14,25 @@ const redirectConfig = {
   permanent: true,
 };
 
-const mapDifficulty = (difficulty) => {
-  switch (difficulty?.toLowerCase()) {
-    case 'junior':
-      return 'easy';
-    case 'semi-senior':
-      return 'intermediate';
-    case 'senior':
-      return 'hard';
-    default:
-      return 'unknown';
-  }
-};
+// const mapDifficulty = (difficulty) => {
+//   switch (difficulty?.toLowerCase()) {
+//     case 'junior':
+//       return 'easy';
+//     case 'semi-senior':
+//       return 'intermediate';
+//     case 'senior':
+//       return 'hard';
+//     default:
+//       return 'unknown';
+//   }
+// };
 
-const getEvents = () => {
-  const data = axios.get(`${BREATHECODE_HOST}/v1/events/all`)
-    .then((res) => res.data)
-    .catch((err) => console.log(err));
-  return data;
-};
+// const getEvents = () => {
+//   const data = axios.get(`${BREATHECODE_HOST}/v1/events/all`)
+//     .then((res) => res.data)
+//     .catch((err) => console.log(err));
+//   return data;
+// };
 
 const getAliasRedirects = async () => {
   const data = axios.get(`${BREATHECODE_HOST}/v1/registry/alias/redirect?academy=${WHITE_LABEL_ACADEMY}`)
@@ -130,24 +131,20 @@ async function generateRedirect() {
   if (!isWhiteLabelAcademy) {
     console.log('Generating redirects...');
 
-    const lessonsList = await getAsset('LESSON,ARTICLE', { exclude_category: 'how-to,como' });
-    const excersisesList = await getAsset('exercise');
-    const projectList = await getAsset('PROJECT').then((data) => data.map((item) => {
-      item.difficulty = mapDifficulty(item.difficulty);
-      return item;
-    }));
-    const howToList = await getAsset('LESSON,ARTICLE').then(
-      (data) => data.filter((l) => l?.category?.slug === 'how-to' || l?.category?.slug === 'como'),
-    );
+    // TODO: adaptar generatedRoutes
+    const lessonsList = await generatedRoutes().then((data) => data.lessons);
+    // ---------------------
 
-    const eventList = await getEvents();
+    console.log('lessonsList:::', lessonsList);
+
     const aliasRedirectList = await getAliasRedirects();
 
     const lessonRedirectList = generateAssetRedirect(lessonsList);
-    const excersisesRedirectList = generateAssetRedirect(excersisesList);
-    const projectRedirectList = generateAssetRedirect(projectList);
-    const howToRedirectList = generateAssetRedirect(howToList);
-    const eventRedirectList = generateAssetRedirect(eventList, 'EVENT');
+    // const excersisesRedirectList = generateAssetRedirect(excersisesList);
+    // const projectRedirectList = generateAssetRedirect(projectList);
+    // const howToRedirectList = generateAssetRedirect(howToList);
+    // const eventRedirectList = generateAssetRedirect(eventList, 'EVENT');
+    const projectList = [];
 
     const aliasRedirectionList = await generateAliasRedirects(aliasRedirectList, projectList)
       .then((redirects) => redirects);
@@ -155,10 +152,10 @@ async function generateRedirect() {
 
     const redirectJson = [
       ...lessonRedirectList,
-      ...excersisesRedirectList,
-      ...projectRedirectList,
-      ...howToRedirectList,
-      ...eventRedirectList,
+      // ...excersisesRedirectList,
+      // ...projectRedirectList,
+      // ...howToRedirectList,
+      // ...eventRedirectList,
     ];
 
     fs.writeFileSync('public/redirects-from-api.json', JSON.stringify(redirectJson, null, 2));
